% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCAMatrix.R
\name{filter}
\alias{filter}
\alias{filters}
\alias{filtering}
\title{Manipulating HCAMatrix filters}
\usage{
filter(x, expr)

filters(x)

filters(x) <- value

\S4method{filter}{HCAMatrix}(x, expr)

\S4method{filters}{HCAMatrix}(x)

\S4method{filters}{HCAMatrix}(x) <- value
}
\arguments{
\item{x}{the object on which to set the filter list
member}

\item{expr}{a filter expression in the form of
the right hand side of a formula, where bare names
(without quotes) are allowed if they are available
fields associated with the HCAMAtrix object, \code{x}}

\item{value}{A list of structured filters (internal use)}
}
\value{
A \code{\link{HCAMatrix}} object with the filter
field replaced by the specified filter expression
}
\description{
Manipulating HCAMatrix filters
}
\note{
Filtering documentation provided by the `GenomicDataCommons` package
}
\section{filter}{

    The \code{filter} is a convenient setter for the filter
    element in \code{\link{HCAMatrix}} objects.
}

\section{filters}{

    The \code{filters} (plural) function is a safe accessor for the filters
    already present in the `HCAMAtrix` API object. The filter can also be
    set using the `filters<-` function setter (advanced use).
}

\examples{
# make an HCAMatrix object to start
hca <- HCAMatrix()

head(available_filters(hca))

hca1 <- filter(hca, genes_detected >= 500)
filters(hca1)

}
