% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCAMatrix.R
\docType{class}
\name{HCAMatrix-class}
\alias{HCAMatrix-class}
\alias{.HCAMatrix}
\title{A class for representing the HCAMatrix API}
\description{
The `HCAMatrix` class is a representation of the `HCAMatrix`
API protocol via OAS version 2.0. The original version OAS 3 was converted
using the APIMatic converter (\url{apimatic.io}).
}
\examples{

HCAMatrix()

}
\seealso{
\link{HCAMatrix}, \link[AnVIL]{Service}
}
