% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHCAMatrix.R
\name{HCAMatrix}
\alias{HCAMatrix}
\alias{getHCAMatrixID}
\alias{loadHCAMatrix}
\title{Obtain expression matrix data from the Human Cell Atlas API service}
\usage{
getHCAMatrixID(bundle_fqids, matrix_query_url = .matrix_query_url,
  verbose = TRUE)

loadHCAMatrix(request_id)
}
\arguments{
\item{bundle_fqids}{A character vector of bundle identifiers}

\item{matrix_query_url}{A single character vector of the API endpoint for
downloading matrix data (defaults to HCA matrix endpoint)}

\item{verbose}{logical (default TRUE) whether to output stepwise messages}

\item{request_id}{A single string obtained from the getHCAMatrixID function}
}
\value{
getHCAMatrixID: A request identifier as a single string

loadHCAMatrix: A LoomExperiment object
}
\description{
Using a vector of data bundle identifiers (`bundle_fqids`), users
can request the associated matrix of expression values. The query submitted
by `getHCAMatrixID` can take some time to be completed. Once the query is
completed, users can use the `request_id` to load the dataset as a
`LoomExperiment` object using the `loadHCAMatrix` function.
}
\details{
The matrix_query_url value points to
    \url{https://matrix.data.humancellatlas.org/v0/matrix}
}
\examples{

bundle_fqids <-
    c("980b814a-b493-4611-8157-c0193590e7d9.2018-11-12T131442.994059Z",
    "7243c745-606d-4827-9fea-65a925d5ab98.2018-11-07T002256.653887Z")

req_id <- getHCAMatrixID(bundle_fqids)

loadHCAMatrix(req_id)

}
