% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mtxzip.R
\name{import_mtxzip}
\alias{import_mtxzip}
\title{Import Human Cell Atlas '.mtx.zip' archives}
\usage{
import_mtxzip(
  path,
  ...,
  exdir = tempfile(),
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path}{character(1) the path to the remote (`http://` or
`https://`) or local `.zip` archive or, when `.data` is
present, the column name (default `"path"`) in which the path
is found.}

\item{...}{additional arguments, not supported.}

\item{exdir}{character(1) directory in which to extract .zip archive.}

\item{overwrite}{logical(1) overwrite existing files in `exdir`?}

\item{verbose}{logical(1) report progress using `message()`.}
}
\value{
`SingleCellExperiment()` representing the expression
    data. Rows represent features and columns cells; counts are
    represented in a sparse matrix.
}
\description{
Import Human Cell Atlas '.mtx.zip' archives
}
\examples{
\dontrun{
    import_mtxzip("path/to/mtx/archive.zip")
}

}
\keyword{internal}
