% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCAMatrix.R
\name{HCAMatrix}
\alias{HCAMatrix}
\title{API Entry function for the Human Cell Atlas Matrix service}
\usage{
HCAMatrix()
}
\value{
An object of class 'HCAMatrix'
}
\description{
This function allows the use of the HCA Matrix API
}
