% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHCAMatrix.R
\name{loadHCAMatrix}
\alias{loadHCAMatrix}
\title{Obtain expression matrix data from the Human Cell Atlas API service}
\usage{
loadHCAMatrix(api, bundle_fqids, verbose = FALSE, names.col = "CellID")
}
\arguments{
\item{api}{An API object of class `HCAMatrix` from the `HCAMatrix`
function}

\item{bundle_fqids}{A character vector of bundle identifiers}

\item{verbose}{logical (default FALSE) whether to output stepwise messages}

\item{names.col}{character (default "CellID") The column name in the
`colData`` metadata to use as column names of the
 \linkS4class{LoomExperiment} object}
}
\value{
A \linkS4class{LoomExperiment} object
}
\description{
Using a vector of data bundle identifiers (`bundle_fqids`),
users can request the associated matrix of expression values. The query
submitted by `loadHCAMatrix` may take some time to be completed. Once the
query is completed, a `LoomExperiment` object is loaded.
}
\details{
The matrix_query_url value points to
    \url{https://matrix.data.humancellatlas.org/v0/matrix}
}
\examples{

hca <- HCAMatrix()

bundle_fqids <-
    c("ffd3bc7b-8f3b-4f97-aa2a-78f9bac93775.2019-05-14T122736.345000Z",
    "f69b288c-fabc-4ac8-b50c-7abcae3731bc.2019-05-14T120110.781000Z",
    "f8ba80a9-71b1-4c15-bcfc-c05a50660898.2019-05-14T122536.545000Z",
    "fd202a54-7085-406d-a92a-aad6dd2d3ef0.2019-05-14T121656.910000Z",
    "fffe55c1-18ed-401b-aa9a-6f64d0b93fec.2019-05-17T233932.932000Z")

loadHCAMatrix(hca, bundle_fqids)

}
